refactor cursor files and functions
